Moving to a somewhat cleaner architecture WIP
