+++ Application/ApiConnections/IApiConnector.cs
@@ -1,0 +1,10 @@
+using Domain.Entities;
+
+namespace Application.ApiConnections
+{
+    public interface IApiConnector
+    {
+        Task<Comment?> GetComment(int id);
+        Task<Comment?> GetLastComment();
+    }
+}
