+++ Application/Interfaces/ICommentsRepository.cs
@@ -1,0 +1,9 @@
+using Domain.Entities;
+
+namespace Application.Interfaces;
+
+public interface ICommentsRepository
+{
+    Task<List<Comment>> GetAllAsync();
+    Task AddAsync(Comment comment);
+}
