+++ Application/Interfaces/IMessage.cs
@@ -1,0 +1,6 @@
+namespace Application.Interfaces;
+
+public interface IMessage
+{
+
+}
