+++ Application/Interfaces/IPublisher.cs
@@ -1,0 +1,6 @@
+namespace Application.Interfaces;
+
+public interface IPublisher
+{
+    void Publish<TMessage>(string exchangeName, TMessage message) //where TMessage : class, IMessage;
+}
