+++ Application/Services/ISorter.cs
@@ -1,0 +1,9 @@
+﻿using Domain.Entities;
+
+namespace Application.Services
+{
+    public interface ISorter
+    {
+        List<Comment> Sort(IEnumerable<Comment> unsorted, SortingParameters parameters);
+    }
+}
