+++ Domain/Entities/Interest.cs
@@ -1,0 +1,8 @@
+namespace Domain.Entities;
+
+public record Interest
+{
+    public string Id { get; init; } = string.Empty;
+
+    public string Text { get; init; } = string.Empty;
+}
