+++ Infrastructure/RabbitConnections/ChannelWrapper.cs
@@ -1,0 +1,32 @@
+using RabbitMQ.Client;
+
+namespace Infrastructure.RabbitConnections;
+
+public class ChannelWrapper
+{
+    private readonly ThreadLocal<ChannelHolder> _channelHolder = new();
+
+    public IModel? Channel
+    {
+        get => _channelHolder.Value?.Context;
+        set
+        {
+            var holder = _channelHolder.Value;
+
+            if (holder is not null)
+            {
+                holder.Context = null;
+            }
+
+            if (value is not null)
+            {
+                _channelHolder.Value = new ChannelHolder { Context = value };
+            }
+        }
+    }
+
+    private class ChannelHolder
+    {
+        public IModel? Context;
+    }
+}
