+++ Infrastructure/RabbitConnections/IChannelFactory.cs
@@ -1,0 +1,8 @@
+using RabbitMQ.Client;
+
+namespace Infrastructure.RabbitConnections;
+
+public interface IChannelFactory
+{
+    IModel? Create();
+}
