--- src/HackerNewsCommentsFeed/ApiConnections/IApiConnector.cs
@@ -1,9 +1,0 @@
-using HackerNewsCommentsFeed.Domain;
-
-namespace HackerNewsCommentsFeed.ApiConnections;
-
-public interface IApiConnector
-{
-    Task<Comment?> GetComment(int id);
-    Task<Comment?> GetLastComment();
-}
