+++ src/HackerNewsCommentsFeed/Controllers/AuthController.cs
@@ -1,6 +1,6 @@
-using HackerNewsCommentsFeed.Utils;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Mvc;
+using Shared.Utils;
 
 namespace HackerNewsCommentsFeed.Controllers;
