--- src/HackerNewsCommentsFeed/Domain/IMessage.cs
@@ -1,6 +1,0 @@
-namespace HackerNewsCommentsFeed.Domain;
-
-public interface IMessage
-{
-    
-}
