--- src/HackerNewsCommentsFeed/RabbitConnections/IChannelFactory.cs
@@ -1,8 +1,0 @@
-using RabbitMQ.Client;
-
-namespace HackerNewsCommentsFeed.RabbitConnections;
-
-public interface IChannelFactory
-{
-    IModel? Create();
-}
