--- src/HackerNewsCommentsFeed/Repositories/ICommentsRepository.cs
@@ -1,9 +1,0 @@
-using HackerNewsCommentsFeed.Domain;
-
-namespace HackerNewsCommentsFeed.Repositories;
-
-public interface ICommentsRepository
-{
-    Task<List<Comment>> GetAllAsync();
-    Task AddAsync(Comment comment);
-}
