+++ src/Tests/UtilsTests.cs
@@ -1,4 +1,6 @@
 ﻿using System.Text.Json;
+using Application.Services;
+using Domain.Entities;
 using FluentAssertions;
 using HackerNewsCommentsFeed.Domain;
 using HackerNewsCommentsFeed.Utils;
